Validacion duration y budget

+++ src/main/kotlin/com/example/evam3/entity/Characters.kt
@@ -1,6 +1,7 @@
 package com.example.evam3.entity
 
 import jakarta.persistence.*
+import java.math.BigDecimal
 
 @Entity
 @Table(name="characters")
@@ -10,7 +11,7 @@ class Characters {
     @Column(updatable = false)
     var id: Long? = null
     var description: String? = null
-    var cost: Double? = null
+    var cost: BigDecimal? = null
     @Column(name="scene_id")
     var sceneId: Long? = null
 }
