+++ src/main/kotlin/com/example/evam3/entity/Film.kt
@@ -1,6 +1,7 @@
 package com.example.evam3.entity
 
 import jakarta.persistence.*
+import java.math.BigInteger
 
 @Entity
 @Table (name="film")
@@ -11,5 +12,5 @@ class Film {
     var id: Long? = null
     var title: String? = null
     var director: String? = null
-    var duration: Long? = null
+    var duration: BigInteger? = null
 }
