+++ src/main/kotlin/com/example/evam3/entity/Scene.kt
@@ -1,6 +1,8 @@
 package com.example.evam3.entity
 
 import jakarta.persistence.*
+import java.math.BigDecimal
+import java.math.BigInteger
 
 @Entity
 @Table(name="scene")
@@ -10,8 +12,8 @@ class Scene {
     @Column(updatable = false)
     var id: Long? = null
     var description: String? = null
-    var budget: Double? = null
-    var minutes: Int? = null
+    var budget: BigDecimal? = null
+    var minutes: BigInteger? = null
     @Column(name="film_id")
     var filmId: Long? = null
 }
